requirements.txt
absl-py==0.12.0
astunparse==1.6.3
attrs==21.2.0
backcall==0.2.0
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.6
clang==5.0
colorama==0.4.4
debugpy==1.5.1
decorator==5.1.0
dill==0.3.4
docopt==0.6.2
entrypoints==0.3
flatbuffers==1.12
future==0.18.2
gast==0.4.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.53.0
grpcio==1.41.0
h5py==3.1.0
idna==3.2
ipykernel==6.5.1
ipython==7.29.0
jedi==0.18.1
jupyter-client==7.1.0
jupyter-core==4.9.1
keras==2.7.0
Keras-Preprocessing==1.1.2
libclang==12.0.0
Markdown==3.3.4
matplotlib-inline==0.1.3
nest-asyncio==1.5.1
numpy==1.19.5
oauthlib==3.1.1
opt-einsum==3.3.0
parso==0.8.2
pickleshare==0.7.5
pipreqs==0.4.11
promise==2.3
prompt-toolkit==3.0.22
protobuf==3.18.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.10.0
python-dateutil==2.8.2
pyzmq==22.3.0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.15.0
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.7.0
tensorflow-datasets==4.4.0
tensorflow-estimator==2.7.0
tensorflow-hub==0.12.0
tensorflow-io-gcs-filesystem==0.22.0
tensorflow-metadata==1.4.0
tensorflowjs==3.9.0
termcolor==1.1.0
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
typing-extensions==3.7.4.3
urllib3==1.26.7
wcwidth==0.2.5
Werkzeug==2.0.2
wrapt==1.12.1
yarg==0.1.9
